requirements.txt
Django==2.2.1
django-cors-headers==3.0.1
django-environ==0.4.5
djangorestframework==3.9.4
psycopg2-binary==2.8.2
requests==2.21.0
responses==0.10.6
pytest-django==3.4.8
